Fixed bugs with test a ROM, added breakpoints, edited readme
