+++ README.md
@@ -1,1 +1,7 @@
-65xx Emulator in C
+Classic65
+
+Description
+Classic65 is a lightweight 6502 and compatibles emulator written in C. Classic65 does not emulate illegal and W65C02 specific opcodes.
+
+Credits
+The emulator was tested using Klaus2m5 test rom. https://github.com/Klaus2m5/6502_65C02_functional_tests
